vm: extract static constructor
