+++ packages/vm/src/index.ts
@@ -1,6 +1,7 @@
 export { Bloom } from './bloom/index.js'
 export { BlockBuilder, BuildStatus } from './buildBlock.js'
 export { buildBlock } from './buildBlock.js'
+export * from './constructors.js'
 export * from './params.js'
 export { encodeReceipt } from './runBlock.js'
 export { runBlock } from './runBlock.js'
